requirements.txt
ftfy
horovod
huggingface_hub
hydra-core==1.3.2
networkx
numpy
omegaconf
open3d
opencv_python
open_clip_torch
pandas
Pillow
plyfile
pyviz3d
regex
safetensors
scikit_learn
timm
tqdm
transformers